backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
chromadb
python-multipart
soundfile
numpy
